SQL Implementation
SQLite DB setup. Parser for visits now inserts info into DB. Generate sample superuser at '/sql/add_superuser'
