+++ TAP/urls.py
@@ -22,5 +22,5 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('parse/', include('parsing.urls')),
     path('', include('pages.urls')),
-    path('sql_test/', include('sql_test.urls'))
+    path('sql/', include('sql.urls'))
 ]
